backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
groq==0.4.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0
PyJWT==2.8.0
